refactoring of accrual, balance and withdrawals updates

+++ internal/entities/user/user.go
@@ -4,6 +4,6 @@ type User struct {
 	UserID   int64    `json:"user_id"`
 	Login    string   `json:"login"`
 	Password Password `json:"-"`
-	Balance  *uint    `json:"balance"`
+	Balance  *float64 `json:"balance"`
 	Token    Token
 }
