+++ internal/service/domain/user/contract.go
@@ -9,5 +9,5 @@ import (
 type UserRepository interface {
 	Create(ctx context.Context, user *userEntity.User) error
 	Find(ctx context.Context, login string) (*userEntity.User, error)
-	Balance(ctx context.Context, UserID int64) (int, error)
+	Balance(ctx context.Context, UserID int64) (float64, error)
 }
