+++ internal/service/domain/user/user.go
@@ -35,7 +35,7 @@ func (u *UserService) Find(ctx context.Context, login string) (*userEntity.User,
 	return user, nil
 }
 
-func (u *UserService) Balance(ctx context.Context, userID int64) (int, error) {
+func (u *UserService) Balance(ctx context.Context, userID int64) (float64, error) {
 	balance, err := u.UserRepository.Balance(ctx, userID)
 	if err != nil {
 		logger.Log.Info("failed to fetch balance", zap.Error(err))
